[로또] step2: LottoMain
