+++ src/main/java/step2/lotto/README.md
@@ -15,7 +15,7 @@
 - [x] \1. 사용자로부터 구매 금액을 입력받는다.
 - [x] \2. 입력한 금액에 해당하는 로또를 발급한다.
 - [x] \3. 발급된 로또를 출력한다.
-- [ ] \4. 당첨 번호를 입력받는다.
-- [ ] \5. 당첨 번호를 기준으로 로또를 등수를 매긴다.
-- [ ] \6. 로또 당첨 통계를 출력한다.
-- [ ] \7. 로또 당첨 통계에 수익률을 출력한다.
+- [x] \4. 당첨 번호를 입력받는다.
+- [x] \5. 당첨 번호를 기준으로 로또를 등수를 매긴다.
+- [x] \6. 로또 당첨 통계를 출력한다.
+- [x] \7. 로또 당첨 통계에 수익률을 출력한다.
