+++ src/test/java/step2/lotto/model/LottoMachineTest.java
@@ -8,6 +8,20 @@ import org.junit.jupiter.api.Test;
 public class LottoMachineTest {
 
     @Test
+    @DisplayName("로또 구입 테스트")
+    void givenPurchaseAmount_whenBuyLotto_thenSuccess() {
+        // given
+        String purchaseAmount = "1000";
+
+        // when
+        LottoMachine LottoMachine = new LottoMachine();
+        LottoTicketList lottoTicketList = LottoMachine.buyLottoTickets(new Money(purchaseAmount));
+
+        // then
+        assertThat(lottoTicketList.value().size()).isEqualTo(1);
+    }
+
+    @Test
     @DisplayName("로또 발행 테스트")
     void givenLottoMachine_whenCreateLottoMachine_thenSuccess() {
         // given
